+++ WMS/Controllers/WebsiteSlidersController.cs
@@ -18,10 +18,6 @@ namespace WMS.Controllers
     public class WebsiteSlidersController : BaseController
     {
         private readonly ITenantWebsiteService _tenantWebsiteService;
-        private readonly IUserService _userService;
-        private readonly IInvoiceService _invoiceService;
-        private readonly ILookupServices _lookupServices;
-        private readonly IMarketServices _marketServices;
         string UploadDirectory = "~/UploadedFiles/WebsiteSliders/";
         string UploadTempDirectory = "~/UploadedFiles/WebsiteSliders/TempFiles/";
         // GET: WebsiteNavigations
@@ -29,16 +25,13 @@ namespace WMS.Controllers
         public WebsiteSlidersController(ICoreOrderService orderService, IMarketServices marketServices, IPropertyService propertyService, IAccountServices accountServices, ILookupServices lookupServices, IUserService userService, IInvoiceService invoiceService, ITenantWebsiteService tenantWebsiteService)
         : base(orderService, propertyService, accountServices, lookupServices)
         {
-            _marketServices = marketServices;
-            _userService = userService;
-            _invoiceService = invoiceService;
-            _lookupServices = lookupServices;
             _tenantWebsiteService = tenantWebsiteService;
         }
 
         // GET: WebsiteSliders
         public ActionResult Index(int SiteId)
         {
+            if (!caSession.AuthoriseSession()) { return Redirect((string)Session["ErrorUrl"]); }
             ViewBag.SiteId = SiteId;
             SiteName(SiteId);
             Session["UploadWebsiteSlider"] = null;
@@ -55,6 +48,7 @@ namespace WMS.Controllers
         // GET: WebsiteSliders/Create
         public ActionResult Create(int SiteId)
         {
+            if (!caSession.AuthoriseSession()) { return Redirect((string)Session["ErrorUrl"]); }
             ViewBag.SiteId = SiteId;
             SiteName(SiteId);
             ViewBag.ControllerName = "WebsiteSliders";
@@ -69,6 +63,7 @@ namespace WMS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create( WebsiteSlider websiteSlider, IEnumerable<DevExpress.Web.UploadedFile> UploadControl)
         {
+            if (!caSession.AuthoriseSession()) { return Redirect((string)Session["ErrorUrl"]); }
             if (ModelState.IsValid)
             {
                 ViewBag.ControllerName = "WebsiteSliders";
@@ -95,6 +90,7 @@ namespace WMS.Controllers
         // GET: WebsiteSliders/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!caSession.AuthoriseSession()) { return Redirect((string)Session["ErrorUrl"]); }
             Session["UploadWebsiteSlider"] = null;
             ViewBag.ControllerName = "WebsiteSliders";
             if (id == null)
@@ -127,6 +123,7 @@ namespace WMS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit( WebsiteSlider websiteSlider, IEnumerable<DevExpress.Web.UploadedFile> UploadControl)
         {
+            if (!caSession.AuthoriseSession()) { return Redirect((string)Session["ErrorUrl"]); }
             if (ModelState.IsValid)
             {
                 string filePath = "";
@@ -153,6 +150,7 @@ namespace WMS.Controllers
         // GET: WebsiteSliders/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!caSession.AuthoriseSession()) { return Redirect((string)Session["ErrorUrl"]); }
             var result = _tenantWebsiteService.RemoveWebsiteSlider((id ?? 0), CurrentUserId);
             return RedirectToAction("Index", new { SiteId = result.SiteID });
         }
